+++ packages/gov-ui-kit/src/modules/components/vote/index.ts
@@ -1,1 +1,2 @@
 export * from './voteDataListItem';
+export * from './voteProposalDataListItem';