+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Implement `VoteProposalDataListItem` with Structure and Skeleton module
+
 ### Changed
 
 -   Remove Radix props dependency